requirements.txt
python-dotenv
requests
urllib3
fastapi
uvicorn
jsonschema